Add border to QR Code on Dark Theme
While trying to scan the QR Code of my wallet address in the wallet, the
dark theme ended up being problematic as my phone was unable to find the
borders of the QR Code. The dark theme's QR Code should have a white
border around it to counteract this problem with scanning when using
dark theme.

I opted for an 8px border and added an element offset margin so it stays
in the place it is supposed to be.