fix: enable sending transactions when no second signature is enabled (#342)
## Proposed changes

Wallets that have no second signature, don't have a second signature field when sending a transaction. As a result, the buttons would not become available as the form could not be validated due to the missing second signature. This PR fixes that issue so wallets without 2nd sig are able to send tx again.

I'm not that familiar with vuelidate, so please let me know if there's a better way to go about this.