fix: do not switch 2 times when selecting a wallet on `WalletSidebar`
This measure avoids a subsequent race-condition that was pushing an additional
entry to the navigation history. The result was that,, when navigating from the
wallet page to the profiles section, the route was ignored.
Since, it was trying to load a wallet page with no address, the router
was redirecting to the wallets section. So the history was:

 1 - /wallets/:address
 2 - /profiles
 3 - /wallets/:address (address = null)