chore(ledger): fix payload encoding
The ARK Ledger App and Ledger-Transport expect bytes when signing a Transaction or Message.

Desktop Wallet passes a string, which shouldn't be legally permitted.
Result of using on a 'dispatch' call?

This was found while writing unit-tests using Ledger's HW Mocking lib.

Next to adding complexity, payloads are also incorrectly-sized.

Even though the resulting chunks are technically valid,
there's no point in the extra 'string' conversions and it wastes ~44-bytes of every Apdu chunk sent (currently 255-bytes/chunk).

You basically lose nearly two-chunks-worth of Apdu calls over the course of a large Transaction (>=255*10).

This PR changes the implementation of Ledger signing to pass a 'Buffer' of bytes vs a 'string'-type.