fix(network-modal): enforce numeric slip44
Changing the Networks Slip44 does not enforce that the value should be only a number.

As a result, a user may assume they can specify their Account index here as well ("111'/1'", "123'/5'", etc).

This results in an invalid Path ("44'/111'/1'/0'/0/0").
The "Slip/Account" is inserted as if it were only the Slip itself, and DW unknowing uses the "Change" index as if it were the "Account" index.
e.g. (m / purpose' / Slip44(coin_type' / account') / change / address_index / nothing)

The proposed change simply enforces that a Slip44 value in the Network Modal only be a number.